♻️ refactor(ai): extract event creation and notification logic to separate function
